Include libuv test case in test-tutorial script
This commit also cleans up the script a bit: It used to sleep for 10
seconds and call subprocess.Popen.communicate() then. This means that a
process could run into a deadlock due to a full pipe buffer, and is
hence explicitly discouraged in the documentation of subprocess. This
commit instead calls communicate() right away and uses a separate thread
to terminate the process after 10 seconds.