Avoid negative values other than -1 in ssize_t / change naming scheme for internal methods
ssize_t technically must only be able to store values from -1 to
SIZE_MAX.

To simplify the building of frames in the process, this commit lets
safer_snprintf (renamed to _rtm_snprintf) and the json encode calls
return pointers to the end of the string instead of the length of the
bytes written. The functions also silently return NULL if they are
passed a NULL pointer. This allows to chain the calls and only check
once after building a frame whether this succeeded or not.

Furthermore, this commit unifies the naming convention for internal
methods: They all have the _rtm-prefix now, making it easy to
distinguish them from other libraries and stdlib calls. It also adds
documentation to these methods.

Finally, the commit fixes some bugs that were too easily fixed to
justify a separate commit.