Reworked JSON parsing
The previous implementation had some undocumented and obscure behaviour,
making it hard to assess whether it handles all corner cases correctly.
Especially it was unclear how it would behave in case of a malicious
mitm. This commit changes the parser logic a bit. While the parser still
isn't capable of validating JSON, it is now clear which cases exactly it
can handle and which it does not.

(Note: To fully validate JSON, the parser would have to match
parentheses, which is possible using the scratch buffer, validate
numbers & other literals.)