Make logging support configurable
Logging takes a lot of space, because it involves lots of string
literals. Embedded platforms don't have that much space, and it doesn't
make much sense to add logging caps anyway. This commit makes logging
output configurable. The API for setting the logger is not removed,
but it won't be invoked if logging is disabled.